==> 02_stanza_preprocess.ipynb <==
import os, re
import pandas as pd
import numpy as np
import time
import stanza

import medspacy, spacy
from medspacy.ner import TargetRule 
from medspacy.visualization import visualize_ent
from medspacy.sentence_splitting import PyRuSHSentencizer

from matplotlib import pyplot as pl

stanza.download('en', package='mimic')
nlp=stanza.Pipeline('en', package='mimic', processors='tokenize')

df = pd.read_csv('NOTEEVENTS.csv', nrows = 500000, low_memory = False)

# Lowercase the column names for easier typing.
df.columns = df.columns.str.lower()

discharge_summaries=df[df.category=='Discharge summary'].text
physician=df[df.category=='Physician '].text
general=df[df.category=='General'].text
consult=df[df.category=='Consult'].text
nursing=df[df.category=='Nursing'].text
respiratory=df[df.category=='Respiratory '].text
rehab=df[df.category=='Rehab Services'].text
nutrition=df[df.category=='Nutrition'].text

start=time.time()
general_sections=[]
for i in range(0,1000):
    doc=nlp(general.iloc[i])
    for sent in doc.sentences:
        fragment=sent.text.strip().replace('\n',' ').replace('  ','')
        if 3<len(fragment.split(' '))<500:
            general_sections.append(fragment)
end=time.time()
print(end-start)

pd.Series(general_sections).iloc[0]

start=time.time()
physician_sections=[]
for i in range(0,100):
    doc=nlp(physician.iloc[i])
    for sent in doc.sentences:
        fragment=sent.text.strip().replace('\n',' ').replace('  ','')
        if 3<len(fragment.split(' '))<500:
            physician_sections.append(fragment)
end=time.time()
print(end-start)

start=time.time()
consult_sections=[]
for i in range(0,min(100, consult.shape[0])):
    doc=nlp(consult.iloc[i])
    for sent in doc.sentences:
        fragment=sent.text.strip().replace('\n',' ').replace('  ','')
        if 3<len(fragment.split(' '))<500:
            consult_sections.append(fragment)
end=time.time()
print(end-start)

start=time.time()
nursing_sections=[]
for i in range(0,min(100, nursing.shape[0])):
    doc=nlp(nursing.iloc[i])
    for sent in doc.sentences:
        fragment=sent.text.strip().replace('\n',' ').replace('  ','')
        if 3<len(fragment.split(' '))<500:
            nursing_sections.append(fragment)
end=time.time()
print(end-start)

start=time.time()
respiratory_sections=[]
for i in range(0,min(100, respiratory.shape[0])):
    doc=nlp(respiratory.iloc[i])
    for sent in doc.sentences:
        fragment=sent.text.strip().replace('\n',' ').replace('  ','')
        if 3<len(fragment.split(' '))<500:
            respiratory_sections.append(fragment)
end=time.time()
print(end-start)

start=time.time()
rehab_sections=[]
for i in range(0,min(100, rehab.shape[0])):
    doc=nlp(rehab.iloc[i])
    for sent in doc.sentences:
        fragment=sent.text.strip().replace('\n',' ').replace('  ','')
        if 3<len(fragment.split(' '))<500:
            rehab_sections.append(fragment)
end=time.time()
print(end-start)

start=time.time()
nutrition_sections=[]
for i in range(0,min(100, nutrition.shape[0])):
    doc=nlp(nutrition.iloc[i])
    for sent in doc.sentences:
        fragment=sent.text.strip().replace('\n',' ').replace('  ','')
        if 3<len(fragment.split(' '))<500:
            nutrition_sections.append(fragment)
end=time.time()
print(end-start)

start=time.time()
dc_sections=[]
for i in range(0,min(100, discharge_summaries.shape[0])):
    doc=nlp(discharge_summaries.iloc[i])
    for sent in doc.sentences:
        fragment=sent.text.strip().replace('\n',' ').replace('  ','')
        if 3<len(fragment.split(' '))<500:
            dc_sections.append(fragment)
end=time.time()
print(end-start)

dc_sums=pd.Series(dc_sections[0:100])
physician_notes=pd.Series(physician_sections[0:100])
general_notes=pd.Series(general_sections[0:100])
consult_notes=pd.Series(consult_sections[0:100])
nursing_notes=pd.Series(nursing_sections[0:100])
resp_notes=pd.Series(respiratory_sections[0:100])
rehab_notes=pd.Series(rehab_sections[0:100])
nutrition_notes=pd.Series(nutrition_sections[0:100])

dc_sums.to_csv('MIMIC_stanza_discharge_summaries.csv', index=False)
physician_notes.to_csv('MIMIC_stanza_physician_notes.csv', index=False)
general_notes.to_csv('MIMIC_stanza_general_notes.csv', index=False)
consult_notes.to_csv('MIMIC_stanza_consult_notes.csv', index=False)
nursing_notes.to_csv('MIMIC_stanza_nursing_notes.csv', index=False)
resp_notes.to_csv('MIMIC_stanza_resp_notes.csv', index=False)
rehab_notes.to_csv('MIMIC_stanza_rehab_notes.csv', index=False)
nutrition_notes.to_csv('MIMIC_stanza_nutrition_notes.csv', index=False)

text_dc=discharge_summaries.iloc[0]
text_phys=physician.iloc[12]
text_consult=consult.iloc[2]

doc=nlp(text_consult)
final_fragments=[]
for sent in doc.sentences:
    fragment=sent.text.strip().replace('\n',' ').replace('  ',' ')
    if 3<len(fragment.split(' '))<500:
        final_fragments.append(fragment)


for frag in final_fragments:
    print(frag)
    print('----'*10)
